ai/book-do-it-ai-agent/pyproject.toml
[project]
name = "book-do-it-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.99.1",
    "python-dotenv>=1.1.1",
]
